pyproject.toml
[project]
name = "backupchan-cli"
version = "0.1.0"
description = "Command-line interface for interacting with Backup-chan."
authors = [
    { name = "Moltony", email = "koronavirusnyj@gmail.com" }
]
dependencies = [
    "backupchan-client-lib",
    "keyring",
    "platformdirs"
]

[project.scripts]
backupchan = "backupchan_cli.main:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
